+++ src/player.h
@@ -21,7 +21,7 @@ struct Player
     uint16_t mechanics_received;//number of boss mechanics received
     uint64_t last_stab_time;  //time stability is going to expire
     uint64_t last_hit_time;       //time player was last hit with a mechanic
-    uint16_t last_mechanic;       //skill id of last failed mechanic
+    uint32_t last_mechanic;       //skill id of last failed mechanic
     bool in_squad;          //currently in squad
 	bool in_combat;
 
@@ -65,7 +65,7 @@ struct Player
     uint64_t getLastHitTime();
     void setLastHitTime(uint64_t new_hit_time);
 
-    uint16_t getLastMechanic();
+    uint32_t getLastMechanic();
     void setLastMechanic(uint32_t new_mechanic);
 
 	uint64_t getLastMechanicHitTime(uint32_t new_mechanic);
